Fixed warnings , improved static-obj-logic and added primitive collision-detection

+++ general/src/database/prepared_statement.h
@@ -55,7 +55,7 @@ namespace database {
             // Return:
             // - number of valid bytes in the result-column
             //______________________________________________________________________________________________________
-            long unsigned int getParamResultLength(const int paramID);
+            long unsigned int getParamResultLength(const size_t paramID);
             //______________________________________________________________________________________________________
             //
             // Description:
